feat(getData2D): Add VP mode for computing D2 and velocity
The changes in this commit:

1. Renamed the executable from `getData2D` to `getData2D-VP`.
2. Added a new file `getData2D-VP.c` that includes the logic for computing the second invariant of the rate-of-deformation tensor (D2) and the velocity field.
3. Removed the old `getData2D.c` file, as it is now replaced by the new `getData2D-VP.c` file.
4. Updated the `VideoAxi.py` file to use the new `getData2D-VP` executable.

These changes were made to add a new mode of operation to the `getData2D` tool, which allows for the computation of the D2 and velocity fields in addition to the conformation tensor. This is a useful feature for analyzing the flow properties in the simulated system.